+++ js/components/graph/__tests__/Node.test.js
@@ -1,11 +1,10 @@
 import React from "react"
-import renderer from "react-test-renderer"
-import { fireEvent } from "@testing-library/react"
+import { fireEvent, act, screen, render } from "@testing-library/react"
 import Node from "../Node"
 import TestGraph from "./TestGraph"
 
 describe("Hybrid Node", () => {
-  it("should match snapshot", () => {
+  it("should match snapshot", async () => {
     const hybridNodeProps = {
       JSON: {
         fill: "#888888",
@@ -53,8 +52,9 @@ describe("Hybrid Node", () => {
       status: "inactive",
       selected: false,
     }
-    const tree = renderer.create(<Node {...hybridNodeProps} />).toJSON()
-    expect(tree).toMatchSnapshot()
+    await act(async () => render(<Node {...hybridNodeProps} />))
+    const nodeElement = screen.getByText("CSC318/418/").closest("g")
+    expect(nodeElement).toMatchSnapshot()
   })
 
   it("should have the 'hybrid' CSS class", async () => {
@@ -115,7 +115,7 @@ describe("Hybrid Node", () => {
 })
 
 describe("Course Node", () => {
-  it("should match snapshot", () => {
+  it("should match snapshot", async () => {
     const courseProps = {
       JSON: {
         fill: "#5dd5b8",
@@ -151,8 +151,9 @@ describe("Course Node", () => {
       status: "takeable",
       selected: false,
     }
-    const tree = renderer.create(<Node {...courseProps} />).toJSON()
-    expect(tree).toMatchSnapshot()
+    await act(async () => render(<Node {...courseProps} />))
+    const nodeElement = screen.getByText("CSC108").closest("g")
+    expect(nodeElement).toMatchSnapshot()
   })
   it("should have the CSS class: 'node'", async () => {
     const graph = await TestGraph.build()